+++ prolif/plotting/complex3d.py
@@ -10,7 +10,8 @@ Plot interactions in 3D --- :mod:`prolif.plotting.complex3d`
 """
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, ClassVar, Dict, Set, Tuple
+from copy import deepcopy
+from typing import TYPE_CHECKING, ClassVar, Dict, Optional, Set, Tuple
 
 import py3Dmol
 from rdkit import Chem
@@ -24,6 +25,7 @@ if TYPE_CHECKING:
     from prolif.fingerprint import Fingerprint
     from prolif.ifp import IFP
     from prolif.molecule import Molecule
+    from prolif.residue import ResidueId
 
 
 class Complex3D:
@@ -108,19 +110,19 @@ class Complex3D:
     INTERACTION_HOVER_CALLBACK: ClassVar[
         str
     ] = """
-    function(atom,viewer) {
-        if(!atom.label) {
-            atom.label = viewer.addLabel(atom.interaction,
-                {position: atom, backgroundColor: 'black', fontColor:'white'});
+    function(shape,viewer) {
+        if(!shape.label) {
+            shape.label = viewer.addLabel(shape.interaction,
+                {position: shape, backgroundColor: 'black', fontColor:'white'});
         }
     }"""
     DISABLE_HOVER_CALLBACK: ClassVar[
         str
     ] = """
-    function(atom,viewer) { 
-        if(atom.label) {
-            viewer.removeLabel(atom.label);
-            delete atom.label;
+    function(obj,viewer) { 
+        if(obj.label) {
+            viewer.removeLabel(obj.label);
+            delete obj.label;
         }
     }"""
 
@@ -165,20 +167,106 @@ class Complex3D:
         centroid = mol.xyz[list(indices)].mean(axis=0)
         return Point3D(*centroid)
 
-    def display(self, size: Tuple[int, int] = (650, 600), display_all: bool = False):
+    def display(
+        self, size: Tuple[int, int] = (650, 600), display_all: bool = False
+    ) -> py3Dmol.view:
         """Display as a py3Dmol widget view.
 
         Parameters
         ----------
         size: Tuple[int, int] = (650, 600)
             The size of the py3Dmol widget view.
-        display_all : bool
+        display_all : bool = False
             Display all occurences for a given pair of residues and interaction, or only
             the shortest one. Not relevant if ``count=False`` in the ``Fingerprint``
             object.
         """
-        v = py3Dmol.view(width=size[0], height=size[1])
+        v = py3Dmol.view(width=size[0], height=size[1], viewergrid=(1, 1), linked=False)
         v.removeAllModels()
+        self._populate_view(v, position=(0, 0), display_all=display_all)
+        return v
+
+    def compare(
+        self,
+        other: Complex3D,
+        *,
+        size: Tuple[int, int] = (900, 600),
+        display_all: bool = False,
+        linked: bool = True,
+        color_unique: Optional[str] = "magentaCarbon",
+    ) -> py3Dmol.view:
+        """Displays the initial complex side-by-side with a second one for easier
+        comparison.
+
+        Parameters
+        ----------
+        other: Complex3D
+            Other ``Complex3D`` object to compare to.
+        size: Tuple[int, int] = (900, 600)
+            The size of the py3Dmol widget view.
+        display_all : bool = False
+            Display all occurences for a given pair of residues and interaction, or only
+            the shortest one. Not relevant if ``count=False`` in the ``Fingerprint``
+            object.
+        linked: bool = True
+            Link mouse interactions (pan, zoom, translate) on both views.
+        color_unique: Optional[str] = "magentaCarbon",
+            Which color to use for residues that have interactions that are found in one
+            complex but not the other. Use ``None`` to disable the color override.
+
+        .. versionadded:: 2.0.1
+
+        """
+        v = py3Dmol.view(
+            width=size[0],
+            height=size[1],
+            linked=linked,
+            viewergrid=(1, 2),
+        )
+        v.removeAllModels()
+
+        # get set of interactions for both poses
+        interactions1 = set(
+            (resid[1], i)
+            for resid, interactions in self.ifp.items()
+            for i in interactions
+        )
+        interactions2 = set(
+            (resid[1], i)
+            for resid, interactions in other.ifp.items()
+            for i in interactions
+        )
+
+        # get residues with interactions specific to pose 1
+        highlights = (
+            {r[0]: color_unique for r in interactions1 - interactions2}
+            if color_unique
+            else {}
+        )
+        self._populate_view(
+            v, position=(0, 0), display_all=display_all, colormap=highlights
+        )
+
+        # get residues with interactions specific to pose 2
+        highlights = (
+            {r[0]: color_unique for r in interactions2 - interactions1}
+            if color_unique
+            else {}
+        )
+        other._populate_view(
+            v, position=(0, 1), display_all=display_all, colormap=highlights
+        )
+        return v
+
+    def _populate_view(
+        self,
+        v: py3Dmol.view,
+        position: Tuple[int, int] = (0, 0),
+        display_all: bool = False,
+        colormap: Optional[Dict[ResidueId, str]] = None,
+    ) -> None:
+        if colormap is None:
+            colormap = {}
 
         models = {}
         mid = -1
@@ -192,8 +280,12 @@ class Complex3D:
             ]:
                 if resid not in models:
                     mid += 1
-                    v.addModel(Chem.MolToMolBlock(res), "sdf")
-                    model = v.getModel()
+                    v.addModel(Chem.MolToMolBlock(res), "sdf", viewer=position)
+                    model = v.getModel(viewer=position)
+                    if resid in colormap:
+                        style = deepcopy(style)
+                        for key in style:
+                            style[key]["colorscheme"] = colormap[resid]
                     model.setStyle({}, style)
                     # add residue label
                     model.setHoverable(
@@ -246,12 +338,13 @@ class Complex3D:
                             "dashed": True,
                             "fromCap": 1,
                             "toCap": 1,
-                        }
+                        },
+                        viewer=position,
                     )
                     # add label when hovering the middle of the dashed line by adding a dummy atom
                     c = Point3D(*get_centroid([p1, p2]))
                     modelID = models[lresid]
-                    model = v.getModel(modelID)
+                    model = v.getModel(modelID, viewer=position)
                     interaction_label = f"{interaction}: {metadata['distance']:.2f}Å"
                     model.addAtoms(
                         [
@@ -278,17 +371,16 @@ class Complex3D:
         # show protein
         mol = Chem.RemoveAllHs(self.prot_mol)
         pdb = Chem.MolToPDBBlock(mol, flavor=0x20 | 0x10)
-        v.addModel(pdb, "pdb")
-        model = v.getModel()
+        v.addModel(pdb, "pdb", viewer=position)
+        model = v.getModel(viewer=position)
         model.setStyle({}, self.PROTEIN_STYLE)
 
         # do the same for ligand if multiple residues
         if self.lig_mol.n_residues >= self.PEPTIDE_THRESHOLD:
             mol = Chem.RemoveAllHs(self.lig_mol)
             pdb = Chem.MolToPDBBlock(mol, flavor=0x20 | 0x10)
-            v.addModel(pdb, "pdb")
-            model = v.getModel()
+            v.addModel(pdb, "pdb", viewer=position)
+            model = v.getModel(viewer=position)
             model.setStyle({}, self.PEPTIDE_STYLE)
 
-        v.zoomTo({"model": list(models.values())})
-        return v
+        v.zoomTo({"model": list(models.values())}, viewer=position)